requirements.txt
Flask==3.1.0
psycopg2_binary==2.9.10
python-dotenv==1.0.1
scikit_learn==1.5.2
pandas==2.2.3
matplotlib==3.9.3
bcrypt==4.3.0
pynput==1.7.7
